Move bulletin saving logic into mixin
